requirements.txt
phidata
python-dotenv
yfinance
duckduckgo-search
packaging 
fastapi 
uvicorn
groq 
openai 
playground